requirements.txt
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
python-dotenv==0.7.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
